convert data module to kmp
