+++ core/data/src/commonMain/kotlin/com/teobaranga/monica/core/data/adapter/UuidSerializer.kt
