+++ core/data/src/commonMain/kotlin/com/teobaranga/monica/core/data/remote/DeleteResponse.kt
